Simplify CLI: remove start command, crate --list, and skills output
- Remove `start` command and its template (start.rs, md/start.md)
- Remove `--list` flag from `crate` command; name is now required
- `crate` command returns only Crate/Version/Source (no skills/guidance)
- Remove dead code from skills.rs (list_output, info_output, CrateAdvice,
  AlwaysSkill, crate_guidance, format_skill_entry, resource helpers)
