+++ src/lib.rs
@@ -1,14 +1,13 @@
 pub mod agents;
 pub mod cli;
 pub mod config;
-pub mod dispatch;
+pub mod crate_command;
 pub mod git_source;
 pub mod hook;
 pub mod hook_schema;
 pub mod init;
 pub mod output;
 pub mod plugins;
-pub mod start;
 pub mod sync;
 
 pub(crate) mod crate_sources;
