+++ src/skills.rs
@@ -12,59 +12,6 @@ use crate::config::Symposium;
 use crate::plugins::{ParsedPlugin, PluginRegistry, SkillGroup};
 use crate::predicate::{self, Predicate};
 
-/// Format the list of skills applicable to workspace crates as display text.
-pub async fn list_output(
-    sym: &Symposium,
-    registry: &PluginRegistry,
-    workspace: &[(String, semver::Version)],
-) -> String {
-    let skills = skills_applicable_to(sym, registry, workspace).await;
-    if skills.is_empty() {
-        "No skills available for crates in the current dependencies.".to_string()
-    } else {
-        let mut out = "Skills available for crates in the current dependencies:\n\n".to_string();
-        for entry in &skills {
-            let crate_names = entry.effective_crate_names();
-            out.push_str(&format_skill_entry(&entry.skill, &crate_names));
-        }
-        out
-    }
-}
-
-/// Fetch crate sources and format info with any matching guidance.
-pub async fn info_output(
-    sym: &Symposium,
-    name: &str,
-    version: Option<&str>,
-    registry: &PluginRegistry,
-    workspace: &[(String, semver::Version)],
-) -> anyhow::Result<String> {
-    let mut fetch = crate::crate_sources::RustCrateFetch::new(name, workspace, sym.cache_dir());
-    if let Some(v) = version {
-        fetch = fetch.version(v);
-    }
-
-    let result = fetch.fetch().await?;
-
-    let mut output = format!(
-        "Crate: {}\nVersion: {}\nSource: {}\n",
-        result.name,
-        result.version,
-        result.path.display()
-    );
-
-    let resolved_version: semver::Version = result
-        .version
-        .parse()
-        .unwrap_or_else(|_| semver::Version::new(0, 0, 0));
-    let advice = crate_guidance(sym, &result.name, &resolved_version, registry).await;
-    if !advice.is_empty() {
-        output.push_str(&advice.format_output());
-    }
-
-    Ok(output)
-}
-
 /// Activation mode for a skill.
 #[derive(Debug, Clone, PartialEq, Default, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -114,72 +61,6 @@ impl Skill {
             pred.collect_crate_names(&mut names);
         }
         names.into_iter().collect()
-    }
-}
-
-/// An always-active skill whose content is inlined in crate output.
-pub struct AlwaysSkill {
-    pub name: String,
-    pub path: PathBuf,
-    pub body: String,
-}
-
-/// Collected advice for a specific crate query.
-pub struct CrateAdvice {
-    /// Content from `activation: always` skills, inlined in output.
-    pub always_skills: Vec<AlwaysSkill>,
-    /// Optional skills with full metadata for agent decision-making.
-    pub optional_skills: Vec<Skill>,
-}
-
-impl CrateAdvice {
-    pub fn is_empty(&self) -> bool {
-        self.always_skills.is_empty() && self.optional_skills.is_empty()
-    }
-
-    /// Format the advice as text to append to crate command output.
-    ///
-    /// Uses `<skill_content>` structured wrapping as recommended by
-    /// https://agentskills.io/client-implementation/adding-skills-support#structured-wrapping
-    pub fn format_output(&self) -> String {
-        let mut out = String::new();
-
-        if !self.always_skills.is_empty() {
-            out.push_str("\n## Guidance\n");
-            for skill in &self.always_skills {
-                let skill_dir = skill.path.parent().unwrap_or(&skill.path);
-                out.push_str(&format!(
-                    "\n<skill_content name=\"{}\">\n\
-                     \n{}\n\
-                     \nSkill directory: {}\n\
-                     Relative paths in this skill are relative to the skill directory.\n",
-                    skill.name,
-                    skill.body,
-                    skill_dir.display()
-                ));
-
-                let resources = list_skill_resources(skill_dir);
-                if !resources.is_empty() {
-                    out.push_str("\n<skill_resources>\n");
-                    for resource in &resources {
-                        out.push_str(&format!("  <file>{resource}</file>\n"));
-                    }
-                    out.push_str("</skill_resources>\n");
-                }
-
-                out.push_str("</skill_content>\n");
-            }
-        }
-
-        if !self.optional_skills.is_empty() {
-            out.push_str("\n## Additional skills available\n\n");
-            for skill in &self.optional_skills {
-                let crate_names = skill.crate_names();
-                out.push_str(&format_skill_entry(skill, &crate_names));
-            }
-        }
-
-        out
     }
 }
 
@@ -238,92 +119,6 @@ pub async fn skills_applicable_to(
     collect_skills_applicable_to(&registry.standalone_skills, &[], for_crates, &mut results);
 
     results
-}
-
-/// Get guidance for a specific crate from installed plugin skills.
-///
-/// This identifies the relevant skills and separates them into
-/// "always" vs "optional" categories. The "always_skills" are
-/// meant to be shown by default.
-async fn crate_guidance(
-    sym: &Symposium,
-    crate_name: &str,
-    crate_version: &semver::Version,
-    registry: &PluginRegistry,
-) -> CrateAdvice {
-    let mut advice = CrateAdvice {
-        always_skills: Vec::new(),
-        optional_skills: Vec::new(),
-    };
-
-    let for_crates = [(crate_name.to_string(), crate_version.clone())];
-    for entry in skills_applicable_to(sym, registry, &for_crates).await {
-        match entry.skill.activation {
-            Activation::Always => {
-                advice.always_skills.push(AlwaysSkill {
-                    name: entry.skill.name().to_string(),
-                    path: entry.skill.path.clone(),
-                    body: entry.skill.body.clone(),
-                });
-            }
-            Activation::Optional => {
-                advice.optional_skills.push(entry.skill);
-            }
-        }
-    }
-
-    advice
-}
-
-/// Format a skill listing entry for display (shared by list and guidance output).
-/// List resource files in a skill directory, as paths relative to that directory.
-///
-/// Excludes `SKILL.md` itself. Returns sorted paths for deterministic output.
-fn list_skill_resources(skill_dir: &Path) -> Vec<String> {
-    let mut resources = Vec::new();
-    collect_resources_recursive(skill_dir, skill_dir, &mut resources);
-    resources.sort();
-    resources
-}
-
-fn collect_resources_recursive(base: &Path, dir: &Path, out: &mut Vec<String>) {
-    let entries = match std::fs::read_dir(dir) {
-        Ok(e) => e,
-        Err(_) => return,
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_dir() {
-            collect_resources_recursive(base, &path, out);
-        } else if path.file_name().is_some_and(|f| f != "SKILL.md") {
-            if let Ok(relative) = path.strip_prefix(base) {
-                out.push(relative.to_string_lossy().into_owned());
-            }
-        }
-    }
-}
-
-fn format_skill_entry(skill: &Skill, crate_names: &[String]) -> String {
-    let mut out = String::new();
-    let name = skill.name();
-    out.push_str(&format!("- **{name}**"));
-    if let Some(desc) = skill.frontmatter.get("description") {
-        out.push_str(&format!(": {desc}"));
-    }
-    out.push('\n');
-    if !crate_names.is_empty() {
-        out.push_str(&format!(
-            "  - Applies to crates: {}\n",
-            crate_names.join(", ")
-        ));
-    }
-    for (key, value) in &skill.frontmatter {
-        if key != "name" && key != "description" && key != "activation" {
-            out.push_str(&format!("  - {key}: {value}\n"));
-        }
-    }
-    out.push_str(&format!("  - Path: {}\n", skill.path.display()));
-    out
 }
 
 /// Discover and load skills for a group, applying pre-fetch filtering.
@@ -1024,79 +819,6 @@ mod tests {
         assert!(results[0].group_crates.is_empty());
     }
 
-    #[tokio::test]
-    async fn guidance_includes_standalone_skills() {
-        use crate::plugins::PluginRegistry;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let skill_dir = tmp.path().join("my-skill");
-        fs::create_dir_all(&skill_dir).unwrap();
-        fs::write(
-            skill_dir.join("SKILL.md"),
-            indoc! {"
-                ---
-                name: standalone-serde
-                description: Standalone serde skill
-                crates: serde
-                activation: always
-                ---
-
-                Use serde standalone.
-            "},
-        )
-        .unwrap();
-
-        let skill = load_standalone_skill(&skill_dir.join("SKILL.md")).unwrap();
-        let registry = PluginRegistry {
-            plugins: Vec::new(),
-            standalone_skills: vec![skill],
-        };
-
-        let sym = crate::config::Symposium::from_dir(tmp.path());
-        let ver = semver::Version::new(1, 0, 0);
-        let advice = crate_guidance(&sym, "serde", &ver, &registry).await;
-        assert_eq!(advice.always_skills.len(), 1);
-        assert_eq!(advice.always_skills[0].name, "standalone-serde");
-        assert!(
-            advice.always_skills[0]
-                .body
-                .contains("Use serde standalone.")
-        );
-    }
-
-    #[tokio::test]
-    async fn guidance_skips_standalone_skill_wrong_crate() {
-        use crate::plugins::PluginRegistry;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let skill_dir = tmp.path().join("my-skill");
-        fs::create_dir_all(&skill_dir).unwrap();
-        fs::write(
-            skill_dir.join("SKILL.md"),
-            indoc! {"
-                ---
-                name: tokio-skill
-                crates: tokio
-                activation: always
-                ---
-
-                Body.
-            "},
-        )
-        .unwrap();
-
-        let skill = load_standalone_skill(&skill_dir.join("SKILL.md")).unwrap();
-        let registry = PluginRegistry {
-            plugins: Vec::new(),
-            standalone_skills: vec![skill],
-        };
-
-        let sym = crate::config::Symposium::from_dir(tmp.path());
-        let ver = semver::Version::new(1, 0, 0);
-        let advice = crate_guidance(&sym, "serde", &ver, &registry).await;
-        assert!(advice.is_empty());
-    }
-
     // --- Discovery ---
 
     #[test]
@@ -1231,71 +953,5 @@ mod tests {
         let defaults = SkillGroup::default();
         let skills = discover_skills(tmp.path(), &defaults);
         assert!(skills.is_empty());
-    }
-
-    // --- CrateAdvice formatting ---
-
-    #[test]
-    fn crate_advice_format_empty() {
-        let advice = CrateAdvice {
-            always_skills: vec![],
-            optional_skills: vec![],
-        };
-        assert!(advice.is_empty());
-        assert_eq!(advice.format_output(), "");
-    }
-
-    #[test]
-    fn crate_advice_format_default_only() {
-        let advice = CrateAdvice {
-            always_skills: vec![AlwaysSkill {
-                name: "skill1".into(),
-                path: PathBuf::from("/skills/serde/SKILL.md"),
-                body: "Use serde this way.".into(),
-            }],
-            optional_skills: vec![],
-        };
-        expect_test::expect![[r#"
-
-            ## Guidance
-
-            <skill_content name="skill1">
-
-            Use serde this way.
-
-            Skill directory: /skills/serde
-            Relative paths in this skill are relative to the skill directory.
-            </skill_content>
-        "#]]
-        .assert_eq(&advice.format_output());
-    }
-
-    #[test]
-    fn crate_advice_format_optional_only() {
-        let advice = CrateAdvice {
-            always_skills: vec![],
-            optional_skills: vec![Skill {
-                frontmatter: BTreeMap::from([
-                    ("name".into(), "adv".into()),
-                    ("description".into(), "Advanced guidance".into()),
-                    ("compatibility".into(), "Requires Python 3.14+".into()),
-                    ("allowed-tools".into(), "Bash(python:*)".into()),
-                ]),
-                crates: vec![],
-                activation: Activation::Optional,
-                body: String::new(),
-                path: PathBuf::from("/path/to/SKILL.md"),
-            }],
-        };
-        expect_test::expect![[r#"
-
-            ## Additional skills available
-
-            - **adv**: Advanced guidance
-              - allowed-tools: Bash(python:*)
-              - compatibility: Requires Python 3.14+
-              - Path: /path/to/SKILL.md
-        "#]]
-        .assert_eq(&advice.format_output());
     }
 }
